requirements.txt
fastapi==0.104.1
mysql-connector-python==8.2.0
bcrypt==4.1.1
python-multipart==0.0.6
passlib==1.7.4
python-jose==3.3.0
uvicorn==0.24.0